zabbix-appliance/ubuntu: fix mysql database creation
When the mysql data directory is empty (e.g. due to mounting in a
container environment to persist the database), the old script
would attempt to call the deprecated mysql_install_db, which also
would fail to find the mysqld binary.
This commit directly calls mysqld now, so that installation from
scratch works again.