Don't create MYSQL_USER unless MYSQL_ROOT_PASSWORD is specified
Here we're setting CREATE_ZBX_DB_USER=true, and rigth below this
test, we're using MYSQL_USER and MYSQL_PASSWORD as root credentials
Then later we create the MYSQL_USER which seems to be a bad idea.

I think it's better to assume if we're not supplying the
MYSQL_ROOT_PASSWORD, the MYSQL_USER/PASSWORD supplied already
exists with the proper permissions.