Added variable "POSTGRES_USE_IMPLICIT_SEARCH_PATH" to skip setting search_path via PGOPTIONS.
Setting the search path via PGOPTIONS prevents connections through pgbouncer,
because setting PGOPTIONS causes psql to set a connection parameter "options",
which is refused by pgbouncer, and "options" is generally unsafe to ignore
as a startup option in pgbouncer.

However, if a user has their postgres users setup with the right search_path,
it is safe to not set the search_path already, because the defaults are right.